refactor:remove_debugging_displays
commented out the various `Display.DebugDraw` as the console was struggling to keep up. The best thing to do is to make the application cash the debug draws and draw them in update. But because this is for debugging and a lack of time, I decided against it. I added a TODO comment instead.